pyproject.toml
[project]
name = "kmsearch"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Shimpei Ueno", email = "" }
]
dependencies = [
    "sqlalchemy>=2.0.22",
    "pandas>=2.0.3",
    "ipykernel>=6.25.2",
    "easyocr>=1.7.1",
    "PyMuPDF>=1.23.5",
    "pillow>=10.1.0",
    "openpyxl>=3.1.2",
    "toml>=0.10.2",
    "lark>=1.1.8",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
